Introduce PVC based repo-host for pgbackrest
With this commit a PVC based backup is possible for pgbackrest.
This can be done by providing Spec.Backup.Pgbackrest.Repo[i].Storage = "pvc"
in the manifest, this will also require to provide Spec.Backup.Pgbackrest.Repo[i].pvcSize = 'xGi'.
When this is option is set, a statefulset with 1 instance is created for the repo-host.
The required global and db section for the pgbackrest configuration file pgbackrest_instance.conf
should be set by the appropriate configmaps for both the postgres pods and the repo-host pod.