Use correct cert locations and permissions
Database pod mounts certificates into /etc/pgbackrest/conf.d.

The code was cross-specifying the certificates. Conceptually one cert
goes on the database pod for both tls-server certificate and TLS
authentication certificate. And the other one goes onto the repo host
for both purposes. Right now it does not matter as they come from the
same secret and the CN is the same. But it might matter when CN is made
to be hostname.