PVC code cert management fixes (#38)
* Copy environment slices to avoid unintended sharing

appendEnv shares memory backing the slices so modifications to one slice
end up overwriting changes done via a different slice object. Need to
explicitly copy the slices to avoid this.

* Make pgbackrest certificate CN = clustername

The previous coding had "cpo-cluster@clustername=*". The portion before
the @ is not adding anything useful so removed it. The =* portion is
pgbackrest config file syntax for which repositories a specified CN has
access to. This should not be in the CN, but only the config file
itself.

* Use correct cert locations and permissions

Database pod mounts certificates into /etc/pgbackrest/conf.d.

The code was cross-specifying the certificates. Conceptually one cert
goes on the database pod for both tls-server certificate and TLS
authentication certificate. And the other one goes onto the repo host
for both purposes. Right now it does not matter as they come from the
same secret and the CN is the same. But it might matter when CN is made
to be hostname.

* Refactor labels handling to include type label everywhere and add type label to KUBERNETES_LABELS

* fixed selector in clusterpods-svc

* changed configmaps for postgres-pod and repo-host

* Add TODO note about fixed cluster domain

* Allow cert usage for both client and server and via both services

* Revert "changed configmaps for postgres-pod and repo-host"

This reverts commit cdabdd993a57409d04c6bcd16e14df2c5b265779.

We want to be using one cert in one host and the other on the other.

* Database pod connects to repo as foo-pgbackrest-repo-host-0.foo-clusterpods.svc.cluster.local

* Add extended key usages to root certificate

---------

Co-authored-by: matthias <matthias@swobblmail.eu>