Add pgbackrest clone support
Adds new attributes to cluster manifest clone section that allows
cloning from pgbackrest sources. Right now the stanza is not
configurable and must be "db". Adds a pgabckrest-clone configmap with
repository configuration that gets mounted to the container.

PVC based clones automatically mount the certificate secret from the
source cluster to be able to clone. Cross namespace clones are currently
not supported.

S3 support is to be tested, but uses same code as backups.